Extract AttachmentLinkHeaderUpdater logic into a worker
