+++ app/workers/asset_manager_attachment_link_header_update_worker.rb
@@ -1,0 +1,20 @@
+class AssetManagerAttachmentLinkHeaderUpdateWorker < WorkerBase
+  def perform(attachment_data)
+    visible_edition = attachment_data.visible_edition_for(nil)
+    return unless visible_edition.present?
+
+    parent_document_url = Whitehall.url_maker.public_document_url(visible_edition)
+
+    enqueue_job(attachment_data.file, parent_document_url)
+    if attachment_data.pdf?
+      enqueue_job(attachment_data.file.thumbnail, parent_document_url)
+    end
+  end
+
+private
+
+  def enqueue_job(uploader, parent_document_url)
+    legacy_url_path = uploader.asset_manager_path
+    AssetManagerUpdateAssetWorker.perform_async(legacy_url_path, parent_document_url: parent_document_url)
+  end
+end
